move updater to package root
